Add AvaloniaUI cross-platform frontend to default template (#13)
* Initial plan

* Add AvaloniaUI frontend template to default template


* Add avalonia android / desktop / browser / iOS

* Create directory if has directory component exists

Path.GetDirectoryName(newPath) can return null (e.g., if newPath has no directory component), which will cause Directory.CreateDirectory(...) to throw. Store the directory name in a variable and only call CreateDirectory when it’s non-empty/non-null.


* InvalidOperationException instead of null reference exception


* Improve avalonia view

* Clean up

* Throw correct exception


---------

+++ Templates/Default/Resources.cs
@@ -1,7 +1,17 @@
 foreach (var path in Directory.GetFiles(Project.ResourcesPath, "*.*", SearchOption.AllDirectories))
 {
 	var newPath = path.Replace(Project.ResourcesPath, Project.OutputPath).Replace("Project.Namespace", Project.Namespace);
-	var dirPath = Path.GetDirectoryName(newPath);
-	Directory.CreateDirectory(dirPath);
-	File.Copy(path, newPath, true);
+	var directoryName = Path.GetDirectoryName(newPath);
+	if (!string.IsNullOrEmpty(directoryName))
+	{
+		Directory.CreateDirectory(directoryName);
+	}
+
+	// Validate modify time of the file, if the file is not modified, skip copying the file to save time
+	var lastWriteTime = File.GetLastWriteTime(path);
+	if (!File.Exists(newPath) || lastWriteTime != File.GetLastWriteTime(newPath))
+	{
+		File.Copy(path, newPath, true);
+		File.SetLastWriteTime(newPath, lastWriteTime);
+	}
 }
